requirements.txt
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0
dspy-ai>=2.3.6
nest-asyncio>=1.6.0
pydantic>=2.0.0
requests>=2.31.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
datasets>=2.15.0
tqdm>=4.65.0
tavily-python>=0.2.6
dspy>=2.3.3 